Lumos/backend/requirements.txt
flask
flask-cors
feedparser
requests
beautifulsoup4
neo4j
dashscope
python-dotenv